Update Steganography.py
Some users may encounter difficulty recognizing images labeled as 'image1' and 'image2'. To address this, I have left a blank space for individuals to input the path of their own images. Additionally, there is a condition implemented: if the sizes of the images provided are not identical, the application will not execute any operations.